Update websockets.php
I force the spliting of handshake request over several packet by setting 152 for $bufferLength.  As expected the handshake data was over severall packet because it size was around ~256.  
 
In the event some client have to send initial handshake over several packet here a fix.

With this fix we buffer the handshake to be process in next call.  

I think we'll need some security check over that to prevent an excessive buffer data if the server never receive what he seek as "\n\n". I guess if after 1024 the server did not receive the end code close the connection.